added generic to provide the ref_clk rate
The generic g_ref_clk_rate was dummy, i.e. never used. The module
pulse_stamper is used with input reference clock (and tm_cycles_i)
of 125MHz and 62.5MHz clock, in the wr_streamers. Added possibility
to define what clock is used (default 125MHz or 62.5MHz). In any
case, the output timestamp is of cycle period of 8ns.